Move AWS file to a package
